backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==2.0.3
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
mango-mcp==1.0.0
python-dotenv==1.0.0
alembic==1.11.1
